remove deprecated marching_cubes_classic implementation
