+++ skimage/measure/__init__.py
@@ -1,6 +1,5 @@
 from ._find_contours import find_contours
-from ._marching_cubes_lewiner import marching_cubes
-from ._marching_cubes_classic import mesh_surface_area
+from ._marching_cubes_lewiner import marching_cubes, mesh_surface_area
 from ._regionprops import (regionprops, perimeter,
                            perimeter_crofton, euler_number, regionprops_table)
 from ._polygon import approximate_polygon, subdivide_polygon
