+++ skimage/measure/tests/test_marching_cubes.py
@@ -11,7 +11,8 @@ def test_marching_cubes_isotropic():
     _, surf = ellipsoid_stats(6, 10, 16)
 
     # Classic
-    verts, faces = marching_cubes(ellipsoid_isotropic, 0., method='_lorensen')
+    verts, faces = marching_cubes(ellipsoid_isotropic, 0.,
+                                  method='lorensen')[:2]
     surf_calc = mesh_surface_area(verts, faces)
     # Test within 1% tolerance for isotropic. Will always underestimate.
     assert surf > surf_calc and surf_calc > surf * 0.99
@@ -32,7 +33,7 @@ def test_marching_cubes_anisotropic():
 
     # Classic
     verts, faces = marching_cubes(ellipsoid_anisotropic, 0.,
-                                  spacing=spacing, method='_lorensen')
+                                  spacing=spacing, method='lorensen')[:2]
     surf_calc = mesh_surface_area(verts, faces)
     # Test within 1.5% tolerance for anisotropic. Will always underestimate.
     assert surf > surf_calc and surf_calc > surf * 0.985
@@ -58,13 +59,13 @@ def test_marching_cubes_anisotropic():
 def test_invalid_input():
     # Classic
     with pytest.raises(ValueError):
-        marching_cubes(np.zeros((2, 2, 1)), 0, method='_lorensen')
+        marching_cubes(np.zeros((2, 2, 1)), 0, method='lorensen')
     with pytest.raises(ValueError):
-        marching_cubes(np.zeros((2, 2, 1)), 1, method='_lorensen')
+        marching_cubes(np.zeros((2, 2, 1)), 1, method='lorensen')
     with pytest.raises(ValueError):
-        marching_cubes(np.ones((3, 3, 3)), 1, spacing=(1, 2), method='_lorensen')
+        marching_cubes(np.ones((3, 3, 3)), 1, spacing=(1, 2), method='lorensen')
     with pytest.raises(ValueError):
-        marching_cubes(np.zeros((20, 20)), 0, method='_lorensen')
+        marching_cubes(np.zeros((20, 20)), 0, method='lorensen')
 
     # Lewiner
     with pytest.raises(ValueError):
@@ -76,23 +77,26 @@ def test_invalid_input():
     with pytest.raises(ValueError):
         marching_cubes(np.zeros((20, 20)), 0)
 
+    # invalid method name
+    ellipsoid_isotropic = ellipsoid(6, 10, 16, levelset=True)
+    with pytest.raises(ValueError):
+        marching_cubes(ellipsoid_isotropic, 0., method='abcd')
+
 
 def test_both_algs_same_result_ellipse():
     # Performing this test on data that does not have ambiguities
 
     sphere_small = ellipsoid(1, 1, 1, levelset=True)
 
-    vertices1, faces1 = marching_cubes(sphere_small, 0, method='_lorensen')[:2]
-    vertices2, faces2 = marching_cubes(sphere_small, 0,
+    vertices1, faces1 = marching_cubes(sphere_small, 0,
                                        allow_degenerate=False)[:2]
-    vertices3, faces3 = marching_cubes(sphere_small, 0,
+    vertices2, faces2 = marching_cubes(sphere_small, 0,
                                        allow_degenerate=False,
                                        method='lorensen')[:2]
 
     # Order is different, best we can do is test equal shape and same
     # vertices present
     assert _same_mesh(vertices1, faces1, vertices2, faces2)
-    assert _same_mesh(vertices1, faces1, vertices3, faces3)
 
 
 def _same_mesh(vertices1, faces1, vertices2, faces2, tol=1e-10):
@@ -133,16 +137,11 @@ def test_both_algs_same_result_donut():
                     64 * ( ((8*y-2)+4)*((8*y-2)+4) + (8*z)**2
                     ) ) + 1025
 
-    vertices1, faces1 = marching_cubes(vol, 0, method='_lorensen')[:2]
+    vertices1, faces1 = marching_cubes(vol, 0, method='lorensen')[:2]
     vertices2, faces2 = marching_cubes(vol, 0)[:2]
-    vertices3, faces3 = marching_cubes(vol, 0, method='lorensen')[:2]
 
     # Old and new alg are different
     assert not _same_mesh(vertices1, faces1, vertices2, faces2)
-    # New classic and new Lewiner are different
-    assert not _same_mesh(vertices2, faces2, vertices3, faces3)
-    # Would have been nice if old and new classic would have been the same
-    # assert _same_mesh(vertices1, faces1, vertices3, faces3, 5)
 
 
 def test_masked_marching_cubes():
@@ -162,13 +161,6 @@ def test_masked_marching_cubes_empty():
     mask = np.array([])
     with pytest.raises(ValueError):
         _ = marching_cubes(ellipsoid_scalar, 0, mask=mask)
-
-
-def test_masked_marching_cubes_old_lewiner():
-    ellipsoid_scalar = ellipsoid(6, 10, 16, levelset=True)
-    mask = np.array([])
-    with pytest.raises(NotImplementedError):
-        _ = marching_cubes(ellipsoid_scalar, 0, mask=mask, method='_lorensen')
 
 
 def test_masked_marching_cubes_all_true():
